requirements.txt
absl-py==2.1.0
attrs==24.2.0
cffi==1.17.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
flatbuffers==24.3.25
fonttools==4.55.0
jax==0.4.35
jaxlib==0.4.35
kiwisolver==1.4.7
matplotlib==3.9.3
mediapipe==0.10.18
ml_dtypes==0.5.0
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
six==1.16.0
sounddevice==0.5.1
wheel==0.45.1
